requirements.txt
flask
flask-cors
openai-whisper
openai
torch>=2.4.0
torchvision
torchaudio
pydantic
python-dotenv
yt-dlp
langchain
langchain-ollama
chromadb
sentence-transformers
langchain-community
